requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors==0.1
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2024.2
